python/requirements.txt
rfc5424-logging-handler==1.4.3